Handle MPI types differently in tracing
Explicitly handle tracing of MPI types (MPI_Datatype, MPI_Comm).
Since its hard (without explicitly mapping each type to a
string) to convert opaque MPI types in a portable way we just
add placeholders for now

This change adds support for tracing with the OpenMPI library.

+++ src/clib/api/spio_put_varm_api.cpp
@@ -14,7 +14,7 @@ int PIOc_put_varm(int ncid, int varid, const PIO_Offset *start, const PIO_Offset
     add_arg("varid", varid).add_arg("*start", start).add_arg("*count", count).
     add_arg("*stride", stride).add_arg("*imap", imap).add_arg("*buf", buf).
     add_arg("bufcount", static_cast<long long int>(bufcount)).
-    add_arg("buftype", static_cast<long long int>(buftype)).flush();
+    add_arg("buftype", buftype).flush();
 #endif
   return PIOc_put_varm_impl(ncid, varid, start, count, stride, imap, buf, bufcount, buftype);
 }
